requirements.txt
transformers>=4.36.0
datasets>=2.2.0
jsonlines
numpy
requests>=2.26.0
scikit_learn>=1.2.0
scipy
sentence_transformers>=2.2.0
tqdm
rich
h5py>=3.8.0
openai>=1.0.0
python-dotenv